Completed day 11
Things learned:
- Controlling the video element. Turned out to be close to the audio element.
- Defining all predefined DOM elements at the beginning makes code more readable.
- Using `this` to refer to the event's initiator is cool, but viewing the function without knowing it's used as a event callback may be confusing if you don't know how this behaves.

At a certain point I got stuck figuring out why my scrubbing/seeking code was acting strange. Then it occured to me I was using e.target.clientWidth to get the full seekbar width. Which is fine as long as it's the seekbar you are clicking and not the highlighted progressBar.